Remove model fix
